perf(webui): load the notification panel on demand
The /chat entry closure carried the whole notification centre, panel markup
included, even though the panel only renders once a reader opens it. That left
the budget with no headroom for anyone else after this branch and main's OOBE
work both claimed it.

The opened panel moves to its own module and loads through `React.lazy`, the way
the command palette already does from the gateway layout. The bell and its
unread dot stay eager because the badge is always on screen. The panel leaves the
entry as a 5.39 kB chunk, /chat measures 223.2 KB, and the ceiling ratchets down
to that rather than keeping the raise this split made unnecessary.

The component tests now settle until the panel mounts instead of asserting on
the render that requested it, since the first open suspends.
